#23 Refacto InputColumn code before implementing cleaning scripts in pyrog client
